* commit '2da3814d5c63a12f969250648cab4d092c3aab72':
  OH-491-fixed-migrations

+++ database/migrations/2020_04_16_190459_add_price_policy_id_to_doctors_table.php
@@ -29,9 +29,11 @@ class AddPricePolicyIdToDoctorsTable extends Migration
      */
     public function down()
     {
-        Schema::table('doctors', function (Blueprint $table) {
-            $table->dropForeign('doctors_price_policy_id_foreign');
-            $table->dropColumn('price_policy_id');
-        });
+        if (Schema::hasColumn('doctors', 'price_policy_id')) {
+            Schema::table('doctors', function (Blueprint $table) {
+                $table->dropForeign('doctors_price_policy_id_foreign');
+                $table->dropColumn('price_policy_id');
+            });
+        }
     }
 }
